refactor(rust): split commands/messages.rs into submodules (types, send, read, ingest, edit_delete, reactions)
