--- pollis-core/src/commands/messages.rs
@@ -1,1802 +1,0 @@
-use rusqlite::OptionalExtension;
-use serde::{Deserialize, Serialize};
-
-use std::sync::Arc;
-use ulid::Ulid;
-
-use crate::error::Result;
-use crate::state::AppState;
-
-use crate::db::queries::MESSAGES_BY_SENDER as QUERY_MESSAGES_BY_SENDER;
-use crate::db::queries::CHANNEL_PREVIEWS as QUERY_CHANNEL_PREVIEWS;
-
-// Envelope cleanup: TTL gate OR watermark gate. Watermark gate is keyed on
-// (user, device) — a multi-device user whose other device hasn't synced keeps
-// envelopes alive until either every device catches up or the TTL expires.
-const CLEANUP_CHANNEL_ENVELOPES: &str = "\
-DELETE FROM message_envelope
- WHERE conversation_id = ?1
-   AND (
-     sent_at < datetime('now', '-30 days')
-     OR sent_at < (
-       SELECT CASE
-                WHEN COUNT(ud.device_id) = COUNT(cw.last_fetched_at)
-                THEN MIN(cw.last_fetched_at)
-                ELSE NULL
-              END
-       FROM group_member gm
-       JOIN channels c ON c.id = ?1 AND c.group_id = gm.group_id
-       JOIN user_device ud ON ud.user_id = gm.user_id
-       LEFT JOIN conversation_watermark cw
-              ON cw.conversation_id = ?1
-             AND cw.user_id = ud.user_id
-             AND cw.device_id = ud.device_id
-     )
-   )";
-
-const CLEANUP_DM_ENVELOPES: &str = "\
-DELETE FROM message_envelope
- WHERE conversation_id = ?1
-   AND (
-     sent_at < datetime('now', '-30 days')
-     OR sent_at < (
-       SELECT CASE
-                WHEN COUNT(ud.device_id) = COUNT(cw.last_fetched_at)
-                THEN MIN(cw.last_fetched_at)
-                ELSE NULL
-              END
-       FROM dm_channel_member dcm
-       JOIN user_device ud ON ud.user_id = dcm.user_id
-       LEFT JOIN conversation_watermark cw
-              ON cw.conversation_id = ?1
-             AND cw.user_id = ud.user_id
-             AND cw.device_id = ud.device_id
-       WHERE dcm.dm_channel_id = ?1
-     )
-   )";
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct Message {
-    pub id: String,
-    pub conversation_id: String,
-    pub sender_id: String,
-    pub content: Option<String>,
-    pub reply_to_id: Option<String>,
-    pub sent_at: String,
-}
-
-/// A message with its group and channel context, used when listing across
-/// all channels (e.g. a user's sent message history).
-#[derive(Debug, Serialize, Deserialize)]
-pub struct MessageWithContext {
-    pub group_id: String,
-    pub group_name: String,
-    pub channel_id: String,
-    pub channel_name: String,
-    pub id: String,
-    pub sender_id: String,
-    pub ciphertext: String,
-    pub sent_at: String,
-}
-
-/// The most recent message in a channel alongside the sender's username,
-/// used to populate channel list previews in the sidebar.
-#[derive(Debug, Serialize, Deserialize)]
-pub struct ChannelPreview {
-    pub group_id: String,
-    pub group_name: String,
-    pub channel_id: String,
-    pub channel_name: String,
-    pub last_message: Option<String>,
-    pub last_sent_at: Option<String>,
-    pub last_sender_id: Option<String>,
-    pub last_sender_username: Option<String>,
-}
-
-pub async fn list_messages(
-    conversation_id: String,
-    limit: Option<i64>,
-    before_id: Option<String>,
-    state: &Arc<AppState>,
-) -> Result<Vec<Message>> {
-    let guard = state.local_db.lock().await;
-    let db = guard.as_ref().ok_or_else(|| crate::error::Error::Other(anyhow::anyhow!("Not signed in")))?;
-    let limit = limit.unwrap_or(50);
-
-    let messages = if let Some(before) = before_id {
-        let before_time: String = db.conn().query_row(
-            "SELECT sent_at FROM message WHERE id = ?1",
-            rusqlite::params![before],
-            |row| row.get(0),
-        ).unwrap_or_else(|_| chrono::Utc::now().to_rfc3339());
-
-        let mut stmt = db.conn().prepare(
-            "SELECT id, conversation_id, sender_id, content, reply_to_id, sent_at
-             FROM message
-             WHERE conversation_id = ?1 AND sent_at < ?2
-             ORDER BY sent_at DESC LIMIT ?3"
-        )?;
-
-        let rows = stmt.query_map(
-            rusqlite::params![conversation_id, before_time, limit],
-            |row| Ok(Message {
-                id: row.get(0)?,
-                conversation_id: row.get(1)?,
-                sender_id: row.get(2)?,
-                content: row.get(3)?,
-                reply_to_id: row.get(4)?,
-                sent_at: row.get(5)?,
-            }),
-        )?;
-
-        rows.filter_map(|r| r.ok()).collect()
-    } else {
-        let mut stmt = db.conn().prepare(
-            "SELECT id, conversation_id, sender_id, content, reply_to_id, sent_at
-             FROM message
-             WHERE conversation_id = ?1
-             ORDER BY sent_at DESC LIMIT ?2"
-        )?;
-
-        let rows = stmt.query_map(
-            rusqlite::params![conversation_id, limit],
-            |row| Ok(Message {
-                id: row.get(0)?,
-                conversation_id: row.get(1)?,
-                sender_id: row.get(2)?,
-                content: row.get(3)?,
-                reply_to_id: row.get(4)?,
-                sent_at: row.get(5)?,
-            }),
-        )?;
-
-        rows.filter_map(|r| r.ok()).collect()
-    };
-
-    Ok(messages)
-}
-
-pub async fn send_message(
-    conversation_id: String,
-    sender_id: String,
-    content: String,
-    reply_to_id: Option<String>,
-    sender_username: Option<String>,
-    state: &Arc<AppState>,
-) -> Result<Message> {
-    state.check_not_outdated()?;
-    let id = Ulid::new().to_string();
-    let now = chrono::Utc::now().to_rfc3339();
-
-    // For group channels, all channels share the group's MLS group (keyed by group_id).
-    // For DM conversations, the MLS group is keyed by conversation_id directly.
-    // is_channel = true means conversation_id is a channel ID; group_id is the LiveKit room name.
-    let (mls_group_id, is_channel) = {
-        let conn = state.remote_db.conn().await?;
-        let mut rows = conn.query(
-            "SELECT group_id FROM channels WHERE id = ?1",
-            libsql::params![conversation_id.clone()],
-        ).await?;
-        match rows.next().await? {
-            Some(row) => (row.get::<String>(0)?, true),
-            None => (conversation_id.clone(), false),
-        }
-    };
-
-    // Block enforcement for DMs: if any other participant in this DM
-    // has a block relationship with the sender (either direction),
-    // silently drop the message. The send appears to succeed to the
-    // sender (message is stored locally so their own history looks
-    // consistent) but it is NOT encrypted, NOT posted to Turso, and
-    // NOT broadcast on LiveKit — the recipient never sees it and no
-    // observable signal reveals the block. Group channels are not
-    // gated here; blocks in groups are purely render-side on the
-    // blocker's client.
-    let suppress_delivery = if !is_channel {
-        let conn = state.remote_db.conn().await?;
-        let mut rows = conn.query(
-            "SELECT user_id
-             FROM dm_channel_member
-             WHERE dm_channel_id = ?1
-               AND user_id <> ?2",
-            libsql::params![conversation_id.clone(), sender_id.clone()],
-        ).await?;
-        let mut blocked = false;
-        while let Some(row) = rows.next().await? {
-            let other: String = row.get(0)?;
-            if crate::commands::blocks::is_blocked_either_way(&conn, &sender_id, &other).await? {
-                blocked = true;
-                break;
-            }
-        }
-        blocked
-    } else {
-        false
-    };
-
-    if suppress_delivery {
-        // Write a local-only row so the sender's own conversation
-        // view stays consistent (history survives reloads). Empty
-        // ciphertext is fine — nothing will ever decrypt this row;
-        // it's only read back via the `content` column.
-        {
-            let guard = state.local_db.lock().await;
-            let db = guard.as_ref().ok_or_else(|| crate::error::Error::Other(
-                anyhow::anyhow!("Not signed in")
-            ))?;
-            let empty: Vec<u8> = Vec::new();
-            db.conn().execute(
-                "INSERT INTO message (id, conversation_id, sender_id, ciphertext, content, reply_to_id, sent_at)
-                 VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7)",
-                rusqlite::params![id, conversation_id, sender_id, empty, content, reply_to_id, now],
-            )?;
-        }
-        return Ok(Message {
-            id,
-            conversation_id,
-            sender_id,
-            content: Some(content),
-            reply_to_id,
-            sent_at: now,
-        });
-    }
-
-    // Poll MLS Welcomes — this device may have been added to the group but
-    // hasn't applied the Welcome yet.
-    {
-        let device_id = state.device_id.lock().await.clone();
-        if let Some(ref did) = device_id {
-            if let Err(e) = crate::commands::mls::poll_mls_welcomes_inner(state, &sender_id, did).await {
-                eprintln!("[messages] send_message: poll_mls_welcomes for {mls_group_id}: {e}");
-            }
-        }
-    }
-
-    // Ensure this device has a local MLS group at the current epoch.
-    // Processes pending commits; falls back to external-join if needed.
-    if let Err(e) = crate::commands::mls::process_pending_commits_inner(state, &mls_group_id, &sender_id).await {
-        eprintln!("[messages] send_message: process_pending_commits for {mls_group_id}: {e}");
-    }
-
-    let ciphertext_remote = {
-        let guard = state.local_db.lock().await;
-        let db = guard.as_ref().ok_or_else(|| crate::error::Error::Other(anyhow::anyhow!("Not signed in")))?;
-
-        let mls_bytes = crate::commands::mls::try_mls_encrypt(db.conn(), &mls_group_id, content.as_bytes())
-            .ok_or_else(|| crate::error::Error::Other(anyhow::anyhow!(
-                "MLS group not initialized for conversation {conversation_id}"
-            )))?;
-
-        let mls_ct_str = format!("mls:{}", hex::encode(&mls_bytes));
-
-        db.conn().execute(
-            "INSERT INTO message (id, conversation_id, sender_id, ciphertext, content, reply_to_id, sent_at)
-             VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7)",
-            rusqlite::params![id, conversation_id, sender_id, mls_bytes, content, reply_to_id, now],
-        )?;
-
-        mls_ct_str
-    };
-
-    // Post to Turso for offline delivery.
-    let conn = state.remote_db.conn().await?;
-    conn.execute(
-        "INSERT INTO message_envelope (id, conversation_id, sender_id, ciphertext, reply_to_id, sent_at)
-         VALUES (?1, ?2, ?3, ?4, ?5, ?6)",
-        libsql::params![id.clone(), conversation_id.clone(), sender_id.clone(), ciphertext_remote, reply_to_id.clone(), now.clone()],
-    ).await?;
-
-    // Notify recipients via LiveKit. Non-fatal — errors are logged, not returned.
-    let uname = sender_username.as_deref();
-    if is_channel {
-        // One LiveKit room per group covers all its channels.
-        // Receivers filter by channel_id in the event payload.
-        if let Err(e) = crate::commands::livekit::publish_new_message_to_room(
-            &state.livekit,
-            &mls_group_id,
-            Some(&conversation_id),
-            None,
-            &sender_id,
-            uname,
-        ).await {
-            eprintln!("[realtime] send_message: publish to group {mls_group_id}: {e}");
-        }
-    } else {
-        // DM: publish directly to the shared DM room (conversation_id is the room name).
-        // Both participants are connected to this room via connect_rooms.
-        if let Err(e) = crate::commands::livekit::publish_new_message_to_room(
-            &state.livekit,
-            &conversation_id,
-            None,
-            Some(&conversation_id),
-            &sender_id,
-            uname,
-        ).await {
-            eprintln!("[realtime] send_message: publish to DM room {conversation_id}: {e}");
-        }
-    }
-
-    Ok(Message {
-        id,
-        conversation_id,
-        sender_id,
-        content: Some(content),
-        reply_to_id,
-        sent_at: now,
-    })
-}
-
-/// A single message row returned by the channel message queries.
-#[derive(Debug, Serialize, Deserialize)]
-pub struct ChannelMessage {
-    pub id: String,
-    pub conversation_id: String,
-    pub sender_id: String,
-    pub sender_username: Option<String>,
-    pub ciphertext: String,
-    pub content: Option<String>,
-    pub reply_to_id: Option<String>,
-    pub sent_at: String,
-    pub edited_at: Option<String>,
-    pub deleted_at: Option<String>,
-}
-
-/// Opaque pagination cursor — the (sent_at, id) of the oldest row on the
-/// current page. Pass it back to fetch the next (older) page.
-#[derive(Debug, Serialize, Deserialize)]
-pub struct MessageCursor {
-    pub sent_at: String,
-    pub id: String,
-}
-
-/// Result of a channel message fetch: the messages (newest-first) and an
-/// optional cursor for fetching the next older page. `next_cursor` is `None`
-/// when fewer than `limit` rows were returned, meaning the beginning of
-/// history has been reached.
-#[derive(Debug, Serialize, Deserialize)]
-pub struct MessagePage {
-    pub messages: Vec<ChannelMessage>,
-    pub next_cursor: Option<MessageCursor>,
-}
-
-/// Pull new envelopes for a channel from remote, decrypt, and persist into the
-/// local `message` table (the authoritative history for this device). Applies
-/// any pending edit envelopes, advances this user's watermark to the max
-/// sent_at successfully persisted up to the first decrypt failure, and runs
-/// envelope GC. Idempotent — repeated calls are a no-op on local state.
-pub async fn ingest_channel_envelopes_inner(
-    state: &Arc<AppState>,
-    user_id: &str,
-    channel_id: &str,
-) -> Result<()> {
-    let conn = state.remote_db.conn().await?;
-
-    // Single round-trip: resolve mls_group_id AND confirm membership. Returns
-    // a row only when the channel exists and the user is a member; otherwise
-    // short-circuit so the read path falls back to whatever is local.
-    let mls_group_id: String = {
-        let mut rows = conn.query(
-            "SELECT c.group_id FROM channels c
-             JOIN group_member gm ON gm.group_id = c.group_id
-             WHERE c.id = ?1 AND gm.user_id = ?2
-             LIMIT 1",
-            libsql::params![channel_id.to_string(), user_id.to_string()],
-        ).await?;
-        match rows.next().await? {
-            Some(row) => row.get::<String>(0)?,
-            None => return Ok(()),
-        }
-    };
-
-    let device_id = state.device_id.lock().await.clone();
-
-    // Advance MLS epoch before decryption so pre-ingest-window commits apply.
-    if let Some(ref did) = device_id {
-        if let Err(e) = crate::commands::mls::poll_mls_welcomes_inner(state, user_id, did).await {
-            eprintln!("[ingest] poll_mls_welcomes for {mls_group_id}: {e}");
-        }
-    }
-    if let Err(e) = crate::commands::mls::process_pending_commits_inner(state, &mls_group_id, user_id).await {
-        eprintln!("[ingest] process_pending_commits for {mls_group_id}: {e}");
-    }
-
-    // Pull only envelopes past this device's watermark. Steady state returns
-    // zero rows. Empty-string COALESCE handles the no-watermark-yet case
-    // (first ingest for this conversation on this device). Ordering by
-    // sent_at ASC is critical so MLS decryption sees epochs in commit order
-    // and the watermark stops at the first decrypt failure.
-    let envelopes: Vec<(String, String, String, Option<String>, Option<String>, String, String)> = {
-        let mut out = Vec::new();
-        let did_param = device_id.clone().unwrap_or_default();
-        let mut rows = conn.query(
-            "SELECT id, sender_id, ciphertext, reply_to_id, target_message_id, sent_at, type
-             FROM message_envelope
-             WHERE conversation_id = ?1
-               AND sent_at > COALESCE(
-                   (SELECT last_fetched_at FROM conversation_watermark
-                    WHERE conversation_id = ?1 AND user_id = ?2 AND device_id = ?3),
-                   ''
-               )
-             ORDER BY sent_at ASC, id ASC",
-            libsql::params![channel_id.to_string(), user_id.to_string(), did_param],
-        ).await?;
-        while let Some(row) = rows.next().await? {
-            out.push((
-                row.get::<String>(0)?,
-                row.get::<String>(1)?,
-                row.get::<String>(2)?,
-                row.get::<Option<String>>(3)?,
-                row.get::<Option<String>>(4)?,
-                row.get::<String>(5)?,
-                row.get::<String>(6)?,
-            ));
-        }
-        out
-    };
-
-    let watermark_ts: Option<String> = persist_envelopes_locally(
-        state,
-        channel_id,
-        &mls_group_id,
-        &envelopes,
-    ).await?;
-
-    if let (Some(ts), Some(did)) = (watermark_ts, device_id.as_ref()) {
-        if let Err(e) = conn.execute(
-            "INSERT INTO conversation_watermark (conversation_id, user_id, device_id, last_fetched_at)
-             VALUES (?1, ?2, ?3, ?4)
-             ON CONFLICT(conversation_id, user_id, device_id) DO UPDATE SET
-               last_fetched_at = MAX(last_fetched_at, excluded.last_fetched_at)",
-            libsql::params![channel_id.to_string(), user_id.to_string(), did.clone(), ts],
-        ).await {
-            eprintln!("[watermark] ingest_channel: upsert failed: {e}");
-        }
-    }
-
-    if let Err(e) = conn.execute(
-        CLEANUP_CHANNEL_ENVELOPES,
-        libsql::params![channel_id.to_string()],
-    ).await {
-        eprintln!("[ingest] channel cleanup failed: {e}");
-    }
-
-    Ok(())
-}
-
-/// Shared envelope-persist loop used by both channel and DM ingest. The
-/// `mls_group_id` differs (channel → group_id, DM → conversation_id) so callers
-/// resolve it and pass it in.
-///
-/// Returns the max sent_at across envelopes that were either successfully
-/// handled (decrypted + stored, or applied as an edit) or already present
-/// locally (idempotent no-op). A decrypt failure does *not* advance the
-/// watermark past its envelope — this keeps failed envelopes alive in
-/// `message_envelope` so a subsequent ingest (after commits/welcomes catch
-/// up) can retry them. The 30-day absolute cutoff in CLEANUP_*_ENVELOPES is
-/// the backstop for envelopes that are permanently undecryptable (e.g.
-/// encrypted to an epoch this device never joined, like pre-join history
-/// for a newly-added member).
-async fn persist_envelopes_locally(
-    state: &Arc<AppState>,
-    conversation_id: &str,
-    mls_group_id: &str,
-    envelopes: &[(String, String, String, Option<String>, Option<String>, String, String)],
-) -> Result<Option<String>> {
-    let guard = state.local_db.lock().await;
-    let db = guard.as_ref().ok_or_else(|| crate::error::Error::Other(anyhow::anyhow!("Not signed in")))?;
-
-    let mut candidate: Option<String> = None;
-    let mut advance = |sent_at: &str| {
-        match candidate.as_deref() {
-            Some(current) if current >= sent_at => {}
-            _ => candidate = Some(sent_at.to_string()),
-        }
-    };
-
-    for (id, sender_id, ciphertext, reply_to_id, target_id, sent_at, env_type) in envelopes {
-        match env_type.as_str() {
-            "message" => {
-                let exists: bool = db.conn().query_row(
-                    "SELECT 1 FROM message WHERE id = ?1",
-                    rusqlite::params![id],
-                    |_| Ok(true),
-                ).optional()?.unwrap_or(false);
-                if exists {
-                    advance(sent_at);
-                    continue;
-                }
-                let ct_bytes = ciphertext.strip_prefix("mls:")
-                    .and_then(|h| hex::decode(h).ok());
-                let plaintext = ct_bytes.as_ref()
-                    .and_then(|b| crate::commands::mls::try_mls_decrypt(db.conn(), mls_group_id, b))
-                    .and_then(|b| String::from_utf8(b).ok());
-                if let (Some(text), Some(bytes)) = (plaintext, ct_bytes) {
-                    let _ = db.conn().execute(
-                        "INSERT OR IGNORE INTO message
-                         (id, conversation_id, sender_id, ciphertext, content, reply_to_id, sent_at)
-                         VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7)",
-                        rusqlite::params![id, conversation_id, sender_id, bytes, text, reply_to_id, sent_at],
-                    );
-                    advance(sent_at);
-                }
-                // Decrypt failed — leave the envelope in message_envelope for
-                // a future retry and do NOT advance the watermark past it.
-            }
-            "edit" => {
-                if let Some(tid) = target_id.as_ref() {
-                    let plaintext = ciphertext.strip_prefix("mls:")
-                        .and_then(|h| hex::decode(h).ok())
-                        .and_then(|b| crate::commands::mls::try_mls_decrypt(db.conn(), mls_group_id, &b))
-                        .and_then(|b| String::from_utf8(b).ok());
-                    if let Some(text) = plaintext {
-                        let now = chrono::Utc::now().to_rfc3339();
-                        let _ = db.conn().execute(
-                            "UPDATE message SET content = ?1, edited_at = ?2
-                             WHERE id = ?3 AND deleted_at IS NULL",
-                            rusqlite::params![text, now, tid],
-                        );
-                        advance(sent_at);
-                    }
-                }
-            }
-            "delete" => {
-                // Admin-issued tombstone for someone else's message. The
-                // ciphertext is empty (no plaintext to decrypt) — the only
-                // payload is target_message_id. Soft-delete the local row so
-                // the read path masks content as "[deleted]".
-                if let Some(tid) = target_id.as_ref() {
-                    let now = chrono::Utc::now().to_rfc3339();
-                    let _ = db.conn().execute(
-                        "UPDATE message SET content = NULL, deleted_at = ?1
-                         WHERE id = ?2 AND deleted_at IS NULL",
-                        rusqlite::params![now, tid],
-                    );
-                    advance(sent_at);
-                }
-            }
-            _ => {
-                advance(sent_at);
-            }
-        }
-    }
-    Ok(candidate)
-}
-
-/// Fetch a page of messages for a channel the user is a member of.
-///
-/// First call: omit `cursor` to get the most recent `limit` messages.
-/// Subsequent calls: pass the `next_cursor` from the previous response to
-/// walk backwards through history. Results are ordered newest-first.
-pub async fn get_channel_messages(
-    user_id: String,
-    channel_id: String,
-    limit: Option<i64>,
-    cursor: Option<MessageCursor>,
-    state: &Arc<AppState>,
-) -> Result<MessagePage> {
-    let limit = limit.unwrap_or(50);
-
-    ingest_channel_envelopes_inner(state, &user_id, &channel_id).await?;
-
-    let mut messages = read_local_channel_page(state, &channel_id, &cursor, limit).await?;
-    attach_sender_usernames(state, &mut messages).await?;
-
-    let next_cursor = if messages.len() == limit as usize {
-        messages.last().map(|m| MessageCursor {
-            sent_at: m.sent_at.clone(),
-            id: m.id.clone(),
-        })
-    } else {
-        None
-    };
-
-    Ok(MessagePage { messages, next_cursor })
-}
-
-/// Read a page of messages for a conversation from the local `message` table,
-/// newest-first. Used by both channel and DM read paths after ingest has
-/// persisted any new envelopes.
-async fn read_local_channel_page(
-    state: &Arc<AppState>,
-    conversation_id: &str,
-    cursor: &Option<MessageCursor>,
-    limit: i64,
-) -> Result<Vec<ChannelMessage>> {
-    let guard = state.local_db.lock().await;
-    let db = guard.as_ref().ok_or_else(|| crate::error::Error::Other(anyhow::anyhow!("Not signed in")))?;
-
-    fn row_to_message(row: &rusqlite::Row<'_>) -> rusqlite::Result<ChannelMessage> {
-        let ct: Vec<u8> = row.get(3)?;
-        let content: Option<String> = row.get(4)?;
-        let deleted_at: Option<String> = row.get(8)?;
-        Ok(ChannelMessage {
-            id: row.get(0)?,
-            conversation_id: row.get(1)?,
-            sender_id: row.get(2)?,
-            sender_username: None,
-            ciphertext: format!("mls:{}", hex::encode(&ct)),
-            // Soft-deleted messages mask content to None regardless of cache.
-            content: if deleted_at.is_some() { None } else { content },
-            reply_to_id: row.get(5)?,
-            sent_at: row.get(6)?,
-            edited_at: row.get(7)?,
-            deleted_at,
-        })
-    }
-
-    let mut rows: Vec<ChannelMessage> = Vec::new();
-    match cursor {
-        None => {
-            let mut stmt = db.conn().prepare(
-                "SELECT id, conversation_id, sender_id, ciphertext, content, reply_to_id, sent_at, edited_at, deleted_at
-                 FROM message
-                 WHERE conversation_id = ?1
-                 ORDER BY sent_at DESC, id DESC
-                 LIMIT ?2"
-            )?;
-            let mapped = stmt.query_map(rusqlite::params![conversation_id, limit], row_to_message)?;
-            for r in mapped {
-                if let Ok(m) = r {
-                    rows.push(m);
-                }
-            }
-        }
-        Some(c) => {
-            let mut stmt = db.conn().prepare(
-                "SELECT id, conversation_id, sender_id, ciphertext, content, reply_to_id, sent_at, edited_at, deleted_at
-                 FROM message
-                 WHERE conversation_id = ?1
-                   AND (sent_at < ?2 OR (sent_at = ?2 AND id < ?3))
-                 ORDER BY sent_at DESC, id DESC
-                 LIMIT ?4"
-            )?;
-            let mapped = stmt.query_map(rusqlite::params![conversation_id, c.sent_at, c.id, limit], row_to_message)?;
-            for r in mapped {
-                if let Ok(m) = r {
-                    rows.push(m);
-                }
-            }
-        }
-    }
-
-    Ok(rows)
-}
-
-/// Batch-resolve sender usernames from the remote `users` table and attach
-/// them to the page. A missing user (deleted, never existed) simply stays as
-/// `None` on that message.
-async fn attach_sender_usernames(
-    state: &Arc<AppState>,
-    messages: &mut [ChannelMessage],
-) -> Result<()> {
-    if messages.is_empty() {
-        return Ok(());
-    }
-    let mut ids: std::collections::HashSet<String> = std::collections::HashSet::new();
-    for m in messages.iter() {
-        ids.insert(m.sender_id.clone());
-    }
-    let ids_vec: Vec<String> = ids.into_iter().collect();
-    let placeholders = (1..=ids_vec.len())
-        .map(|i| format!("?{i}"))
-        .collect::<Vec<_>>()
-        .join(",");
-    let sql = format!("SELECT id, username FROM users WHERE id IN ({placeholders})");
-    let params: Vec<libsql::Value> = ids_vec
-        .iter()
-        .map(|s| libsql::Value::Text(s.clone()))
-        .collect();
-
-    let conn = state.remote_db.conn().await?;
-    let mut rows = conn.query(&sql, libsql::params_from_iter(params)).await?;
-    let mut map: std::collections::HashMap<String, String> = std::collections::HashMap::new();
-    while let Some(row) = rows.next().await? {
-        let id: String = row.get(0)?;
-        let name: String = row.get(1)?;
-        map.insert(id, name);
-    }
-    for m in messages.iter_mut() {
-        m.sender_username = map.get(&m.sender_id).cloned();
-    }
-    Ok(())
-}
-
-/// Attach sender usernames from the local `user_cache` table; for any
-/// sender_ids missing from the cache, do one batched remote fetch and
-/// write the results back. After the first read of a channel/DM, the
-/// cache is warm and subsequent reads are zero-remote.
-async fn attach_sender_usernames_local(
-    state: &Arc<AppState>,
-    messages: &mut [ChannelMessage],
-) -> Result<()> {
-    if messages.is_empty() {
-        return Ok(());
-    }
-
-    let mut ids: std::collections::HashSet<String> = std::collections::HashSet::new();
-    for m in messages.iter() {
-        ids.insert(m.sender_id.clone());
-    }
-    let ids_vec: Vec<String> = ids.into_iter().collect();
-
-    let mut found: std::collections::HashMap<String, String> = std::collections::HashMap::new();
-    let missing: Vec<String> = {
-        let guard = state.local_db.lock().await;
-        let db = guard.as_ref().ok_or_else(|| crate::error::Error::Other(anyhow::anyhow!("Not signed in")))?;
-        let placeholders = (1..=ids_vec.len())
-            .map(|i| format!("?{i}"))
-            .collect::<Vec<_>>()
-            .join(",");
-        let sql = format!("SELECT id, username FROM user_cache WHERE id IN ({placeholders})");
-        let mut stmt = db.conn().prepare(&sql)?;
-        let mapped = stmt.query_map(rusqlite::params_from_iter(ids_vec.iter()), |row| {
-            Ok((row.get::<_, String>(0)?, row.get::<_, String>(1)?))
-        })?;
-        for r in mapped {
-            if let Ok((id, name)) = r {
-                found.insert(id, name);
-            }
-        }
-        ids_vec.iter().filter(|i| !found.contains_key(*i)).cloned().collect()
-    };
-
-    if !missing.is_empty() {
-        let placeholders = (1..=missing.len())
-            .map(|i| format!("?{i}"))
-            .collect::<Vec<_>>()
-            .join(",");
-        let sql = format!("SELECT id, username FROM users WHERE id IN ({placeholders})");
-        let params: Vec<libsql::Value> = missing
-            .iter()
-            .map(|s| libsql::Value::Text(s.clone()))
-            .collect();
-        let conn = state.remote_db.conn().await?;
-        match conn.query(&sql, libsql::params_from_iter(params)).await {
-            Ok(mut rows) => {
-                let mut fetched: Vec<(String, String)> = Vec::new();
-                while let Some(row) = rows.next().await? {
-                    let id: String = row.get(0)?;
-                    let name: String = row.get(1)?;
-                    fetched.push((id.clone(), name.clone()));
-                    found.insert(id, name);
-                }
-                drop(rows);
-                if !fetched.is_empty() {
-                    let guard = state.local_db.lock().await;
-                    if let Some(db) = guard.as_ref() {
-                        for (id, name) in &fetched {
-                            let _ = db.conn().execute(
-                                "INSERT INTO user_cache (id, username, updated_at) VALUES (?1, ?2, datetime('now'))
-                                 ON CONFLICT(id) DO UPDATE SET username = ?2, updated_at = datetime('now')",
-                                rusqlite::params![id, name],
-                            );
-                        }
-                    }
-                }
-            }
-            Err(e) => {
-                // Offline or transient — leave the missing names as None.
-                // Next ingest / read while online will fill them in.
-                eprintln!("[messages] attach_sender_usernames_local: remote fallback failed: {e}");
-            }
-        }
-    }
-
-    for m in messages.iter_mut() {
-        m.sender_username = found.get(&m.sender_id).cloned();
-    }
-    Ok(())
-}
-
-/// Local-only read of a channel page. Does NOT trigger ingest — callers
-/// fire `ingest_channel_envelopes` separately, off the critical render
-/// path. Usernames come from the local `user_cache` (with a one-shot
-/// remote fallback for cache misses, see `attach_sender_usernames_local`).
-pub async fn read_channel_messages(
-    channel_id: String,
-    limit: Option<i64>,
-    cursor: Option<MessageCursor>,
-    state: &Arc<AppState>,
-) -> Result<MessagePage> {
-    let limit = limit.unwrap_or(50);
-    let mut messages = read_local_channel_page(state, &channel_id, &cursor, limit).await?;
-    attach_sender_usernames_local(state, &mut messages).await?;
-
-    let next_cursor = if messages.len() == limit as usize {
-        messages.last().map(|m| MessageCursor {
-            sent_at: m.sent_at.clone(),
-            id: m.id.clone(),
-        })
-    } else {
-        None
-    };
-
-    Ok(MessagePage { messages, next_cursor })
-}
-
-/// Local-only read of a DM page. Mirrors `read_channel_messages`.
-pub async fn read_dm_messages(
-    dm_channel_id: String,
-    limit: Option<i64>,
-    cursor: Option<MessageCursor>,
-    state: &Arc<AppState>,
-) -> Result<MessagePage> {
-    let limit = limit.unwrap_or(50);
-    let mut messages = read_local_channel_page(state, &dm_channel_id, &cursor, limit).await?;
-    attach_sender_usernames_local(state, &mut messages).await?;
-
-    let next_cursor = if messages.len() == limit as usize {
-        messages.last().map(|m| MessageCursor {
-            sent_at: m.sent_at.clone(),
-            id: m.id.clone(),
-        })
-    } else {
-        None
-    };
-
-    Ok(MessagePage { messages, next_cursor })
-}
-
-/// Frontend-triggerable ingest for a channel. Used by LiveKit real-time hints
-/// and channel-focus pre-warm paths that want to persist new envelopes without
-/// reading a page.
-pub async fn ingest_channel_envelopes(
-    user_id: String,
-    channel_id: String,
-    state: &Arc<AppState>,
-) -> Result<()> {
-    ingest_channel_envelopes_inner(state, &user_id, &channel_id).await
-}
-
-/// All messages sent by a given user across all their channels,
-/// ordered by group name, then channel name, then timestamp.
-pub async fn list_messages_by_sender(
-    sender_id: String,
-    state: &Arc<AppState>,
-) -> Result<Vec<MessageWithContext>> {
-    let conn = state.remote_db.conn().await?;
-
-    let mut rows = conn.query(QUERY_MESSAGES_BY_SENDER, libsql::params![sender_id]).await?;
-
-    let mut messages = Vec::new();
-    while let Some(row) = rows.next().await? {
-        messages.push(MessageWithContext {
-            group_id: row.get(0)?,
-            group_name: row.get(1)?,
-            channel_id: row.get(2)?,
-            channel_name: row.get(3)?,
-            id: row.get(4)?,
-            sender_id: row.get(5)?,
-            ciphertext: row.get(6)?,
-            sent_at: row.get(7)?,
-        });
-    }
-
-    Ok(messages)
-}
-
-/// Last message and sender username for every channel the given user belongs to,
-/// ordered most-recently-active first. Channels with no messages appear last.
-pub async fn list_channel_previews(
-    user_id: String,
-    state: &Arc<AppState>,
-) -> Result<Vec<ChannelPreview>> {
-    let conn = state.remote_db.conn().await?;
-
-    let mut rows = conn.query(QUERY_CHANNEL_PREVIEWS, libsql::params![user_id]).await?;
-
-    let mut previews = Vec::new();
-    while let Some(row) = rows.next().await? {
-        previews.push(ChannelPreview {
-            group_id: row.get(0)?,
-            group_name: row.get(1)?,
-            channel_id: row.get(2)?,
-            channel_name: row.get(3)?,
-            last_message: row.get(4)?,
-            last_sent_at: row.get(5)?,
-            last_sender_id: row.get(6)?,
-            last_sender_username: row.get(7)?,
-        });
-    }
-
-    Ok(previews)
-}
-
-/// Pull new envelopes for a DM from remote, decrypt, and persist into the local
-/// `message` table. DM MLS group is keyed by conversation_id directly.
-pub async fn ingest_dm_envelopes_inner(
-    state: &Arc<AppState>,
-    user_id: &str,
-    dm_channel_id: &str,
-) -> Result<()> {
-    let conn = state.remote_db.conn().await?;
-
-    let is_member: bool = {
-        let mut rows = conn.query(
-            "SELECT 1 FROM dm_channel_member
-             WHERE dm_channel_id = ?1 AND user_id = ?2
-             LIMIT 1",
-            libsql::params![dm_channel_id.to_string(), user_id.to_string()],
-        ).await?;
-        rows.next().await?.is_some()
-    };
-    if !is_member {
-        return Ok(());
-    }
-
-    let device_id = state.device_id.lock().await.clone();
-
-    if let Some(ref did) = device_id {
-        if let Err(e) = crate::commands::mls::poll_mls_welcomes_inner(state, user_id, did).await {
-            eprintln!("[ingest] poll_mls_welcomes for DM {dm_channel_id}: {e}");
-        }
-    }
-    if let Err(e) = crate::commands::mls::process_pending_commits_inner(state, dm_channel_id, user_id).await {
-        eprintln!("[ingest] process_pending_commits for DM {dm_channel_id}: {e}");
-    }
-
-    // Watermark-filtered envelope read — see ingest_channel_envelopes_inner.
-    let envelopes: Vec<(String, String, String, Option<String>, Option<String>, String, String)> = {
-        let mut out = Vec::new();
-        let did_param = device_id.clone().unwrap_or_default();
-        let mut rows = conn.query(
-            "SELECT id, sender_id, ciphertext, reply_to_id, target_message_id, sent_at, type
-             FROM message_envelope
-             WHERE conversation_id = ?1
-               AND sent_at > COALESCE(
-                   (SELECT last_fetched_at FROM conversation_watermark
-                    WHERE conversation_id = ?1 AND user_id = ?2 AND device_id = ?3),
-                   ''
-               )
-             ORDER BY sent_at ASC, id ASC",
-            libsql::params![dm_channel_id.to_string(), user_id.to_string(), did_param],
-        ).await?;
-        while let Some(row) = rows.next().await? {
-            out.push((
-                row.get::<String>(0)?,
-                row.get::<String>(1)?,
-                row.get::<String>(2)?,
-                row.get::<Option<String>>(3)?,
-                row.get::<Option<String>>(4)?,
-                row.get::<String>(5)?,
-                row.get::<String>(6)?,
-            ));
-        }
-        out
-    };
-
-    let watermark_ts: Option<String> = persist_envelopes_locally(
-        state,
-        dm_channel_id,
-        dm_channel_id,
-        &envelopes,
-    ).await?;
-
-    if let (Some(ts), Some(did)) = (watermark_ts, device_id.as_ref()) {
-        if let Err(e) = conn.execute(
-            "INSERT INTO conversation_watermark (conversation_id, user_id, device_id, last_fetched_at)
-             VALUES (?1, ?2, ?3, ?4)
-             ON CONFLICT(conversation_id, user_id, device_id) DO UPDATE SET
-               last_fetched_at = MAX(last_fetched_at, excluded.last_fetched_at)",
-            libsql::params![dm_channel_id.to_string(), user_id.to_string(), did.clone(), ts],
-        ).await {
-            eprintln!("[watermark] ingest_dm: upsert failed: {e}");
-        }
-    }
-
-    if let Err(e) = conn.execute(
-        CLEANUP_DM_ENVELOPES,
-        libsql::params![dm_channel_id.to_string()],
-    ).await {
-        eprintln!("[ingest] dm cleanup failed: {e}");
-    }
-
-    Ok(())
-}
-
-/// Fetch a page of messages for a DM channel the user is a member of.
-/// Results are ordered newest-first.
-pub async fn get_dm_messages(
-    user_id: String,
-    dm_channel_id: String,
-    limit: Option<i64>,
-    cursor: Option<MessageCursor>,
-    state: &Arc<AppState>,
-) -> Result<MessagePage> {
-    let limit = limit.unwrap_or(50);
-
-    ingest_dm_envelopes_inner(state, &user_id, &dm_channel_id).await?;
-
-    let mut messages = read_local_channel_page(state, &dm_channel_id, &cursor, limit).await?;
-    attach_sender_usernames(state, &mut messages).await?;
-
-    let next_cursor = if messages.len() == limit as usize {
-        messages.last().map(|m| MessageCursor {
-            sent_at: m.sent_at.clone(),
-            id: m.id.clone(),
-        })
-    } else {
-        None
-    };
-
-    Ok(MessagePage { messages, next_cursor })
-}
-
-/// Frontend-triggerable ingest for a DM channel.
-pub async fn ingest_dm_envelopes(
-    user_id: String,
-    dm_channel_id: String,
-    state: &Arc<AppState>,
-) -> Result<()> {
-    ingest_dm_envelopes_inner(state, &user_id, &dm_channel_id).await
-}
-
-/// A search result from the local message cache.
-#[derive(Debug, Serialize, Deserialize)]
-pub struct SearchResult {
-    pub message_id: String,
-    pub conversation_id: String,
-    pub sender_id: String,
-    pub content: String,
-    pub sent_at: String,
-    /// Surrounding context — same as content for now.
-    pub snippet: String,
-}
-
-/// Search the local plaintext message cache using a LIKE query.
-/// Only messages where content IS NOT NULL are searched (i.e. decrypted messages).
-/// Results are ordered newest-first.
-pub async fn search_messages(
-    query: String,
-    limit: Option<i64>,
-    state: &Arc<AppState>,
-) -> Result<Vec<SearchResult>> {
-    let guard = state.local_db.lock().await;
-    let db = guard.as_ref().ok_or_else(|| crate::error::Error::Other(anyhow::anyhow!("Not signed in")))?;
-    let limit = limit.unwrap_or(50);
-    let pattern = format!("%{}%", query);
-
-    let mut stmt = db.conn().prepare(
-        "SELECT id, conversation_id, sender_id, content, sent_at
-         FROM message
-         WHERE content IS NOT NULL AND content LIKE ?1
-         ORDER BY sent_at DESC LIMIT ?2"
-    )?;
-
-    let rows = stmt.query_map(
-        rusqlite::params![pattern, limit],
-        |row| {
-            let content: String = row.get(3)?;
-            let snippet = content.clone();
-            Ok(SearchResult {
-                message_id: row.get(0)?,
-                conversation_id: row.get(1)?,
-                sender_id: row.get(2)?,
-                content,
-                sent_at: row.get(4)?,
-                snippet,
-            })
-        },
-    )?;
-
-    Ok(rows.filter_map(|r| r.ok()).collect())
-}
-
-/// Delete a message.
-///
-/// Two paths, selected automatically by comparing `user_id` (the caller) to the
-/// stored sender of the target message:
-///
-/// **Self-delete** (caller is the original sender): removes the envelope from
-/// Turso (preventing future delivery to anyone who hasn't fetched it yet) and
-/// removes the row from the sender's local message cache. Recipients who
-/// already received the message keep it — no retroactive broadcast.
-///
-/// **Admin-delete** (caller is a different user, must be a group admin in the
-/// channel's group): writes a `type='delete'` tombstone envelope to Turso so
-/// every other member soft-deletes the message on their next ingest, also
-/// removes the original message envelope and any pending edit, soft-deletes
-/// the admin's own local row, and broadcasts a `deleted_message` realtime
-/// event so currently-connected clients invalidate their cache immediately.
-/// Admin-delete is rejected for DM messages (no admin concept in 1:1 DMs).
-///
-/// Attachment cleanup: if the message had one or more attachments, each
-/// content_hash is reference-counted against the caller's other non-deleted
-/// local messages. When no other local message references the same hash, the
-/// `attachment_object` row is removed from Turso and the R2 object is deleted.
-/// R2 deletion is best-effort and must not fail the overall delete; orphaned
-/// R2 objects can be reclaimed by a future sweep.
-pub async fn delete_message(
-    message_id: String,
-    user_id: String,
-    state: &Arc<AppState>,
-) -> Result<()> {
-    let conn = state.remote_db.conn().await?;
-
-    // Resolve the message's original sender + conversation. Prefer the remote
-    // envelope (authoritative across devices) but fall back to the local row
-    // if Turso has already cleaned up the envelope by watermark/TTL.
-    let (msg_sender_id, conversation_id): (String, String) = {
-        let mut rows = conn.query(
-            "SELECT sender_id, conversation_id FROM message_envelope
-             WHERE id = ?1 AND type = 'message'",
-            libsql::params![message_id.clone()],
-        ).await?;
-        if let Some(row) = rows.next().await? {
-            (row.get::<String>(0)?, row.get::<String>(1)?)
-        } else {
-            let guard = state.local_db.lock().await;
-            let db = guard.as_ref().ok_or_else(|| crate::error::Error::Other(anyhow::anyhow!("Not signed in")))?;
-            let local: Option<(String, String)> = db.conn()
-                .query_row(
-                    "SELECT sender_id, conversation_id FROM message WHERE id = ?1",
-                    rusqlite::params![message_id],
-                    |row| Ok((row.get::<_, String>(0)?, row.get::<_, String>(1)?)),
-                )
-                .optional()?;
-            match local {
-                Some(t) => t,
-                None => {
-                    return Err(crate::error::Error::Other(anyhow::anyhow!(
-                        "Message not found"
-                    )));
-                }
-            }
-        }
-    };
-
-    let is_admin_delete = msg_sender_id != user_id;
-
-    if is_admin_delete {
-        // Admin path: caller must be an admin in the group that owns this
-        // channel. DMs (no `channels` row) are not moderatable.
-        let group_id: String = {
-            let mut rows = conn.query(
-                "SELECT group_id FROM channels WHERE id = ?1",
-                libsql::params![conversation_id.clone()],
-            ).await?;
-            match rows.next().await? {
-                Some(row) => row.get(0)?,
-                None => {
-                    return Err(crate::error::Error::Other(anyhow::anyhow!(
-                        "only the sender can delete this message"
-                    )));
-                }
-            }
-        };
-
-        let mut role_rows = conn.query(
-            "SELECT role FROM group_member WHERE group_id = ?1 AND user_id = ?2",
-            libsql::params![group_id.clone(), user_id.clone()],
-        ).await?;
-        let role: String = if let Some(row) = role_rows.next().await? {
-            row.get(0)?
-        } else {
-            return Err(crate::error::Error::Other(anyhow::anyhow!(
-                "only the sender can delete this message"
-            )));
-        };
-        if role != "admin" {
-            return Err(crate::error::Error::Other(anyhow::anyhow!(
-                "only group admins can delete other members' messages"
-            )));
-        }
-
-        // Remove the original message envelope and any pending edit so
-        // late-joiners or unsynced devices never receive the now-deleted
-        // content. Then write the tombstone envelope so every existing
-        // member soft-deletes on next ingest.
-        conn.execute(
-            "DELETE FROM message_envelope WHERE id = ?1",
-            libsql::params![message_id.clone()],
-        ).await?;
-        conn.execute(
-            "DELETE FROM message_envelope WHERE target_message_id = ?1 AND type = 'edit'",
-            libsql::params![message_id.clone()],
-        ).await?;
-
-        let tombstone_id = Ulid::new().to_string();
-        let now = chrono::Utc::now().to_rfc3339();
-        conn.execute(
-            "INSERT INTO message_envelope
-                 (id, conversation_id, sender_id, ciphertext, sent_at, type, target_message_id)
-             VALUES (?1, ?2, ?3, '', ?4, 'delete', ?5)",
-            libsql::params![tombstone_id, conversation_id.clone(), user_id.clone(), now.clone(), message_id.clone()],
-        ).await?;
-
-        // Soft-delete locally and collect orphaned attachments. The admin
-        // may not have a local row (joined after the message was sent and
-        // it's already aged out) — that's fine, the tombstone in Turso is
-        // what propagates the delete.
-        let orphaned: Vec<AttachmentRef> = {
-            let guard = state.local_db.lock().await;
-            let db = guard.as_ref().ok_or_else(|| crate::error::Error::Other(anyhow::anyhow!("Not signed in")))?;
-
-            let content: Option<String> = db.conn()
-                .query_row(
-                    "SELECT content FROM message WHERE id = ?1",
-                    rusqlite::params![message_id],
-                    |row| row.get(0),
-                )
-                .optional()?;
-
-            db.conn().execute(
-                "UPDATE message SET content = NULL, deleted_at = ?1
-                 WHERE id = ?2 AND deleted_at IS NULL",
-                rusqlite::params![now, message_id],
-            )?;
-
-            let raw = content.unwrap_or_default();
-            let attachments = parse_attachment_refs(&raw);
-            if attachments.is_empty() {
-                Vec::new()
-            } else {
-                // Reference-count against every non-deleted local message,
-                // not just the admin's own — if any local copy still
-                // references the hash (e.g. another member also attached
-                // the same file in a separate message they sent), keep R2.
-                filter_orphaned_locally_all(db.conn(), &attachments)?
-            }
-        };
-
-        for att in orphaned {
-            cleanup_attachment(&state, &att).await;
-        }
-
-        // Broadcast so currently-connected clients invalidate their message
-        // cache without waiting for a refetch. Non-fatal — ingest of the
-        // tombstone envelope is the durable path.
-        if let Err(e) = crate::commands::livekit::publish_deleted_message_to_room(
-            &state.livekit,
-            &group_id,
-            Some(&conversation_id),
-            None,
-            &user_id,
-            &message_id,
-        ).await {
-            eprintln!("[realtime] delete_message: publish to group {group_id}: {e}");
-        }
-
-        return Ok(());
-    }
-
-    // Self-delete path (caller is the original sender).
-    //
-    // Remove the message envelope. Best-effort — may already be gone.
-    conn.execute(
-        "DELETE FROM message_envelope WHERE id = ?1 AND sender_id = ?2",
-        libsql::params![message_id.clone(), user_id.clone()],
-    ).await?;
-
-    // Remove any pending edit envelope for this message.
-    conn.execute(
-        "DELETE FROM message_envelope WHERE target_message_id = ?1 AND type = 'edit'",
-        libsql::params![message_id.clone()],
-    ).await?;
-
-    // Read the local plaintext content before deleting so we can inspect any
-    // embedded attachment metadata. Then delete the local row and compute
-    // which attachments are no longer referenced by any of this user's other
-    // non-deleted messages. Done inside a single lock scope to avoid races
-    // with concurrent sends.
-    let orphaned: Vec<AttachmentRef> = {
-        let guard = state.local_db.lock().await;
-        let db = guard.as_ref().ok_or_else(|| crate::error::Error::Other(anyhow::anyhow!("Not signed in")))?;
-
-        let content: Option<String> = db.conn()
-            .query_row(
-                "SELECT content FROM message WHERE id = ?1 AND sender_id = ?2",
-                rusqlite::params![message_id, user_id],
-                |row| row.get(0),
-            )
-            .optional()?;
-
-        let rows_affected = db.conn().execute(
-            "DELETE FROM message WHERE id = ?1 AND sender_id = ?2",
-            rusqlite::params![message_id, user_id],
-        )?;
-
-        if rows_affected == 0 {
-            return Err(crate::error::Error::Other(anyhow::anyhow!(
-                "Message not found or you are not the sender"
-            )));
-        }
-
-        let raw = match content {
-            Some(r) => r,
-            None => String::new(),
-        };
-        let attachments = parse_attachment_refs(&raw);
-        if attachments.is_empty() {
-            Vec::new()
-        } else {
-            filter_orphaned_locally(db.conn(), &user_id, &attachments)?
-        }
-    };
-
-    for att in orphaned {
-        cleanup_attachment(&state, &att).await;
-    }
-
-    Ok(())
-}
-
-/// Attachment identifier extracted from a message's plaintext JSON payload.
-#[derive(Debug, Clone)]
-struct AttachmentRef {
-    content_hash: String,
-    r2_key: String,
-}
-
-/// Parse the `_att` array out of a message's local plaintext content and
-/// return the (content_hash, r2_key) pairs. Returns an empty Vec for plain
-/// text messages, malformed JSON, or any missing fields.
-fn parse_attachment_refs(raw: &str) -> Vec<AttachmentRef> {
-    if !raw.starts_with('{') {
-        return Vec::new();
-    }
-    let Ok(parsed) = serde_json::from_str::<serde_json::Value>(raw) else {
-        return Vec::new();
-    };
-    let Some(atts) = parsed.get("_att").and_then(|v| v.as_array()) else {
-        return Vec::new();
-    };
-    atts.iter()
-        .filter_map(|a| {
-            let hash = a.get("hash")?.as_str()?.to_string();
-            let key = a.get("key")?.as_str()?.to_string();
-            if hash.is_empty() || key.is_empty() {
-                return None;
-            }
-            Some(AttachmentRef { content_hash: hash, r2_key: key })
-        })
-        .collect()
-}
-
-/// Return the subset of the given attachments that are not referenced by any
-/// of the user's other non-deleted local messages. Scans the sender's local
-/// message cache only — cross-user references are invisible because
-/// attachment metadata lives inside the MLS-encrypted payload.
-fn filter_orphaned_locally(
-    conn: &rusqlite::Connection,
-    user_id: &str,
-    candidates: &[AttachmentRef],
-) -> Result<Vec<AttachmentRef>> {
-    let mut stmt = conn.prepare(
-        "SELECT content FROM message
-         WHERE sender_id = ?1 AND deleted_at IS NULL AND content IS NOT NULL",
-    )?;
-    let rows = stmt.query_map(rusqlite::params![user_id], |row| row.get::<_, String>(0))?;
-
-    let mut still_referenced = std::collections::HashSet::<String>::new();
-    for row in rows {
-        let content = row?;
-        for att in parse_attachment_refs(&content) {
-            still_referenced.insert(att.content_hash);
-        }
-    }
-
-    Ok(candidates
-        .iter()
-        .filter(|a| !still_referenced.contains(&a.content_hash))
-        .cloned()
-        .collect())
-}
-
-/// Same as `filter_orphaned_locally` but scans every non-deleted local
-/// message regardless of sender. Used by admin-delete: the admin may
-/// themselves have re-sent the same attachment, or someone else's message
-/// in the admin's local cache may still reference the same hash. Cross-user
-/// references on remote devices remain invisible (convergent encryption
-/// re-creates the dedup row on re-upload, so a future re-share is safe).
-fn filter_orphaned_locally_all(
-    conn: &rusqlite::Connection,
-    candidates: &[AttachmentRef],
-) -> Result<Vec<AttachmentRef>> {
-    let mut stmt = conn.prepare(
-        "SELECT content FROM message
-         WHERE deleted_at IS NULL AND content IS NOT NULL",
-    )?;
-    let rows = stmt.query_map([], |row| row.get::<_, String>(0))?;
-
-    let mut still_referenced = std::collections::HashSet::<String>::new();
-    for row in rows {
-        let content = row?;
-        for att in parse_attachment_refs(&content) {
-            still_referenced.insert(att.content_hash);
-        }
-    }
-
-    Ok(candidates
-        .iter()
-        .filter(|a| !still_referenced.contains(&a.content_hash))
-        .cloned()
-        .collect())
-}
-
-/// Delete an attachment's Turso dedup row and its R2 object. Best-effort on
-/// both: failures are logged, never bubbled. The attachment_object row is
-/// removed first — if R2 deletion fails, a future re-upload will re-register
-/// the row and overwrite the object, restoring a consistent state.
-async fn cleanup_attachment(state: &AppState, att: &AttachmentRef) {
-    let remote_result = async {
-        let conn = state.remote_db.conn().await?;
-        conn.execute(
-            "DELETE FROM attachment_object WHERE content_hash = ?1",
-            libsql::params![att.content_hash.clone()],
-        ).await?;
-        Ok::<_, crate::error::Error>(())
-    }
-    .await;
-
-    if let Err(e) = remote_result {
-        eprintln!(
-            "[delete_message] failed to remove attachment_object for {}: {e}",
-            att.content_hash
-        );
-    }
-
-    if let Err(e) = crate::commands::r2::delete_r2_object(state, &att.r2_key).await {
-        eprintln!(
-            "[delete_message] failed to delete R2 object {} (hash {}): {e}",
-            att.r2_key, att.content_hash
-        );
-    }
-}
-
-/// Edit a message: updates the sender's local plaintext cache immediately and
-/// publishes an encrypted edit envelope to Turso so all other group members
-/// receive the updated content on their next fetch.
-///
-/// The edit envelope uses type='edit' with target_message_id pointing at the
-/// original message. A DELETE+INSERT replaces any prior pending edit, so Turso
-/// never holds more than one edit per message per conversation.
-pub async fn edit_message(
-    conversation_id: String,
-    message_id: String,
-    user_id: String,
-    new_content: String,
-    state: &Arc<AppState>,
-) -> Result<()> {
-    let envelope_id = Ulid::new().to_string();
-    let now = chrono::Utc::now().to_rfc3339();
-
-    // Resolve the MLS group for this conversation (channel → group_id, DM → conversation_id).
-    let mls_group_id = {
-        let conn = state.remote_db.conn().await?;
-        let mut rows = conn.query(
-            "SELECT group_id FROM channels WHERE id = ?1",
-            libsql::params![conversation_id.clone()],
-        ).await?;
-        match rows.next().await? {
-            Some(row) => row.get::<String>(0)?,
-            None => conversation_id.clone(),
-        }
-    };
-
-    // Encrypt the new content with MLS and update local cache atomically.
-    // First attempt — if the group is missing (e.g. local DB was wiped),
-    // transparently repair and retry.
-    let needs_repair = {
-        let guard = state.local_db.lock().await;
-        let db = guard.as_ref().ok_or_else(|| crate::error::Error::Other(anyhow::anyhow!("Not signed in")))?;
-        crate::commands::mls::try_mls_encrypt(db.conn(), &mls_group_id, new_content.as_bytes()).is_none()
-    };
-
-    if needs_repair {
-        crate::commands::mls::repair_mls_group(state, &mls_group_id, &user_id).await?;
-    }
-
-    let ciphertext_remote = {
-        let guard = state.local_db.lock().await;
-        let db = guard.as_ref().ok_or_else(|| crate::error::Error::Other(anyhow::anyhow!("Not signed in")))?;
-
-        let mls_bytes = crate::commands::mls::try_mls_encrypt(db.conn(), &mls_group_id, new_content.as_bytes())
-            .ok_or_else(|| crate::error::Error::Other(anyhow::anyhow!(
-                "MLS group not initialized for conversation {conversation_id}"
-            )))?;
-
-        let rows_affected = db.conn().execute(
-            "UPDATE message SET content = ?1, edited_at = ?2
-             WHERE id = ?3 AND sender_id = ?4 AND deleted_at IS NULL",
-            rusqlite::params![new_content, now, message_id, user_id],
-        )?;
-
-        if rows_affected == 0 {
-            return Err(crate::error::Error::Other(anyhow::anyhow!(
-                "Message not found, already deleted, or you are not the sender"
-            )));
-        }
-
-        format!("mls:{}", hex::encode(&mls_bytes))
-    };
-
-    // Replace any existing edit envelope for this message with the new one.
-    // DELETE + INSERT rather than ON CONFLICT to stay compatible with older libsql.
-    let conn = state.remote_db.conn().await?;
-    conn.execute(
-        "DELETE FROM message_envelope
-         WHERE conversation_id = ?1 AND target_message_id = ?2 AND type = 'edit'",
-        libsql::params![conversation_id.clone(), message_id.clone()],
-    ).await?;
-
-    conn.execute(
-        "INSERT INTO message_envelope
-             (id, conversation_id, sender_id, ciphertext, sent_at, type, target_message_id)
-         VALUES (?1, ?2, ?3, ?4, ?5, 'edit', ?6)",
-        libsql::params![envelope_id, conversation_id.clone(), user_id.clone(), ciphertext_remote, now, message_id.clone()],
-    ).await?;
-
-    // Notify recipients via LiveKit so they invalidate their cache immediately.
-    // Non-fatal — errors are logged, not returned.
-    let is_channel = mls_group_id != conversation_id;
-    let room_id = mls_group_id;
-    if is_channel {
-        if let Err(e) = crate::commands::livekit::publish_edited_message_to_room(
-            &state.livekit,
-            &room_id,
-            Some(&conversation_id),
-            None,
-            &user_id,
-            &message_id,
-        ).await {
-            eprintln!("[realtime] edit_message: publish to group {room_id}: {e}");
-        }
-    } else {
-        if let Err(e) = crate::commands::livekit::publish_edited_message_to_room(
-            &state.livekit,
-            &room_id,
-            None,
-            Some(&conversation_id),
-            &user_id,
-            &message_id,
-        ).await {
-            eprintln!("[realtime] edit_message: publish to DM room {room_id}: {e}");
-        }
-    }
-
-    Ok(())
-}
-
-/// Aggregated emoji reaction for a message.
-/// `user_ids` is the list of users who reacted with this emoji.
-#[derive(Debug, Serialize, Deserialize)]
-pub struct Reaction {
-    pub emoji: String,
-    pub user_ids: Vec<String>,
-    pub count: u32,
-}
-
-/// Add an emoji reaction to a message.
-/// Silently succeeds if the reaction already exists (UNIQUE constraint).
-pub async fn add_reaction(
-    message_id: String,
-    user_id: String,
-    emoji: String,
-    state: &Arc<AppState>,
-) -> Result<()> {
-    let conn = state.remote_db.conn().await?;
-    let id = Ulid::new().to_string();
-    let now = chrono::Utc::now().to_rfc3339();
-
-    conn.execute(
-        "INSERT OR IGNORE INTO message_reaction (id, message_id, user_id, emoji, created_at)
-         VALUES (?1, ?2, ?3, ?4, ?5)",
-        libsql::params![id, message_id, user_id, emoji, now],
-    ).await?;
-
-    Ok(())
-}
-
-/// Remove an emoji reaction from a message.
-/// Silently succeeds if the reaction does not exist.
-pub async fn remove_reaction(
-    message_id: String,
-    user_id: String,
-    emoji: String,
-    state: &Arc<AppState>,
-) -> Result<()> {
-    let conn = state.remote_db.conn().await?;
-
-    conn.execute(
-        "DELETE FROM message_reaction WHERE message_id = ?1 AND user_id = ?2 AND emoji = ?3",
-        libsql::params![message_id, user_id, emoji],
-    ).await?;
-
-    Ok(())
-}
-
-/// Get all reactions for a message, grouped by emoji.
-pub async fn get_reactions(
-    message_id: String,
-    state: &Arc<AppState>,
-) -> Result<Vec<Reaction>> {
-    let conn = state.remote_db.conn().await?;
-
-    let mut rows = conn.query(
-        "SELECT emoji, user_id FROM message_reaction WHERE message_id = ?1 ORDER BY created_at ASC",
-        libsql::params![message_id],
-    ).await?;
-
-    // Collect all (emoji, user_id) rows and group by emoji.
-    let mut grouped: std::collections::HashMap<String, Vec<String>> = std::collections::HashMap::new();
-    while let Some(row) = rows.next().await? {
-        let emoji: String = row.get(0)?;
-        let uid: String = row.get(1)?;
-        grouped.entry(emoji).or_default().push(uid);
-    }
-
-    let reactions: Vec<Reaction> = grouped
-        .into_iter()
-        .map(|(emoji, user_ids)| {
-            let count = user_ids.len() as u32;
-            Reaction { emoji, user_ids, count }
-        })
-        .collect();
-
-    Ok(reactions)
-}
-
-#[cfg(test)]
-mod tests {
-    use rusqlite::Connection;
-
-    use crate::db::BASELINE_SQL as BASELINE;
-
-    // Both queries operate on the remote schema. Tests use rusqlite in-memory
-    // (same SQLite dialect, no libsql threading conflict in test binaries).
-    fn db() -> Connection {
-        let conn = Connection::open_in_memory().unwrap();
-        conn.execute_batch("PRAGMA foreign_keys=ON;").unwrap();
-        conn.execute_batch(BASELINE).unwrap();
-        conn
-    }
-
-    fn setup(conn: &Connection) {
-        // Users
-        conn.execute("INSERT INTO users (id, email, username) VALUES ('alice', 'alice@x.com', 'alice')", []).unwrap();
-        conn.execute("INSERT INTO users (id, email, username) VALUES ('bob',   'bob@x.com',   'bob')", []).unwrap();
-
-        // Groups (alphabetical names so ordering is deterministic)
-        conn.execute("INSERT INTO groups (id, name, owner_id) VALUES ('g-personal', 'personal', 'alice')", []).unwrap();
-        conn.execute("INSERT INTO groups (id, name, owner_id) VALUES ('g-work',     'work',     'alice')", []).unwrap();
-
-        // Both users are members of both groups
-        for gid in ["g-personal", "g-work"] {
-            for uid in ["alice", "bob"] {
-                conn.execute(
-                    "INSERT INTO group_member (group_id, user_id) VALUES (?1, ?2)",
-                    rusqlite::params![gid, uid],
-                ).unwrap();
-            }
-        }
-
-        // Channels (alphabetical names within groups)
-        conn.execute("INSERT INTO channels (id, group_id, name, created_at) VALUES ('ch-personal-random',  'g-personal', 'random',      '2024-01-01T00:00:00Z')", []).unwrap();
-        conn.execute("INSERT INTO channels (id, group_id, name, created_at) VALUES ('ch-work-engineering', 'g-work',     'engineering', '2024-01-01T00:00:00Z')", []).unwrap();
-        conn.execute("INSERT INTO channels (id, group_id, name, created_at) VALUES ('ch-work-general',     'g-work',     'general',     '2024-01-01T00:00:00Z')", []).unwrap();
-
-        // Messages — alice sends in all three channels, bob sends once in work/general
-        conn.execute("INSERT INTO message_envelope (id, conversation_id, sender_id, ciphertext, sent_at) VALUES ('m1', 'ch-work-general',     'alice', 'hello team',  '2024-01-01T10:00:00Z')", []).unwrap();
-        conn.execute("INSERT INTO message_envelope (id, conversation_id, sender_id, ciphertext, sent_at) VALUES ('m2', 'ch-work-general',     'bob',   'hi alice',    '2024-01-01T10:01:00Z')", []).unwrap();
-        conn.execute("INSERT INTO message_envelope (id, conversation_id, sender_id, ciphertext, sent_at) VALUES ('m3', 'ch-work-engineering', 'alice', 'ship it',     '2024-01-02T09:00:00Z')", []).unwrap();
-        conn.execute("INSERT INTO message_envelope (id, conversation_id, sender_id, ciphertext, sent_at) VALUES ('m4', 'ch-personal-random',  'alice', 'lol',         '2024-01-03T12:00:00Z')", []).unwrap();
-        conn.execute("INSERT INTO message_envelope (id, conversation_id, sender_id, ciphertext, sent_at) VALUES ('m5', 'ch-work-general',     'alice', 'see you all', '2024-01-04T17:00:00Z')", []).unwrap();
-    }
-
-    #[test]
-    fn messages_by_sender_ordered_by_group_then_channel_then_time() {
-        let conn = db();
-        setup(&conn);
-
-        let mut stmt = conn.prepare(super::QUERY_MESSAGES_BY_SENDER).unwrap();
-        // (group_name, channel_name, sent_at)
-        let results: Vec<(String, String, String)> = stmt.query_map(
-            rusqlite::params!["alice"],
-            |row| Ok((row.get(1)?, row.get(3)?, row.get(7)?)),
-        ).unwrap().map(|r| r.unwrap()).collect();
-
-        // alice sent 4 messages: m1, m3, m4, m5 (not m2 which is bob's)
-        // Expected order: personal/random, work/engineering, work/general (x2 by time)
-        assert_eq!(results.len(), 4);
-        assert_eq!(results[0], ("personal".into(), "random".into(),      "2024-01-03T12:00:00Z".into()));
-        assert_eq!(results[1], ("work".into(),     "engineering".into(), "2024-01-02T09:00:00Z".into()));
-        assert_eq!(results[2], ("work".into(),     "general".into(),     "2024-01-01T10:00:00Z".into()));
-        assert_eq!(results[3], ("work".into(),     "general".into(),     "2024-01-04T17:00:00Z".into()));
-    }
-
-    #[test]
-    fn messages_by_sender_excludes_other_senders() {
-        let conn = db();
-        setup(&conn);
-
-        let mut stmt = conn.prepare(super::QUERY_MESSAGES_BY_SENDER).unwrap();
-        let count = stmt.query_map(rusqlite::params!["bob"], |_| Ok(()))
-            .unwrap().count();
-
-        // Bob only sent m2
-        assert_eq!(count, 1);
-    }
-
-    #[test]
-    fn channel_previews_ordered_most_recent_first() {
-        let conn = db();
-        setup(&conn);
-
-        let mut stmt = conn.prepare(super::QUERY_CHANNEL_PREVIEWS).unwrap();
-        // (channel_id, last_message, last_sender_username)
-        let results: Vec<(String, Option<String>, Option<String>)> = stmt.query_map(
-            rusqlite::params!["alice"],
-            |row| Ok((row.get(2)?, row.get(4)?, row.get(7)?)),
-        ).unwrap().map(|r| r.unwrap()).collect();
-
-        assert_eq!(results.len(), 3, "alice belongs to 3 channels");
-
-        // Most recent activity:
-        //   work/general     — m5 at 2024-01-04 (sender: alice)
-        //   personal/random  — m4 at 2024-01-03 (sender: alice)
-        //   work/engineering — m3 at 2024-01-02 (sender: alice)
-        let ids: Vec<&str> = results.iter().map(|(id, _, _)| id.as_str()).collect();
-        assert_eq!(ids, ["ch-work-general", "ch-personal-random", "ch-work-engineering"]);
-
-        let (_, msg, sender) = &results[0];
-        assert_eq!(msg.as_deref(), Some("see you all"));
-        assert_eq!(sender.as_deref(), Some("alice"));
-    }
-
-    #[test]
-    fn channel_previews_last_message_is_most_recent_not_first() {
-        let conn = db();
-        setup(&conn);
-
-        // work/general has m1 (alice), m2 (bob), m5 (alice) — preview should show m5
-        let mut stmt = conn.prepare(super::QUERY_CHANNEL_PREVIEWS).unwrap();
-        let results: Vec<(String, Option<String>, Option<String>)> = stmt.query_map(
-            rusqlite::params!["bob"],
-            |row| Ok((row.get(2)?, row.get(4)?, row.get(7)?)),
-        ).unwrap().map(|r| r.unwrap()).collect();
-
-        let general = results.iter().find(|(id, _, _)| id == "ch-work-general").unwrap();
-        assert_eq!(general.1.as_deref(), Some("see you all"), "preview should show m5 not m1 or m2");
-        assert_eq!(general.2.as_deref(), Some("alice"), "sender of last message is alice");
-    }
-
-    #[test]
-    fn channel_previews_empty_channel_appears_last() {
-        let conn = db();
-        setup(&conn);
-
-        // Add an empty channel both users are in (via g-work membership)
-        conn.execute("INSERT INTO channels (id, group_id, name, created_at) VALUES ('ch-work-quiet', 'g-work', 'quiet', '2024-01-01T00:00:00Z')", []).unwrap();
-
-        let mut stmt = conn.prepare(super::QUERY_CHANNEL_PREVIEWS).unwrap();
-        let results: Vec<(String, Option<String>)> = stmt.query_map(
-            rusqlite::params!["bob"],
-            |row| Ok((row.get(2)?, row.get(4)?)),
-        ).unwrap().map(|r| r.unwrap()).collect();
-
-        let last = results.last().unwrap();
-        assert_eq!(last.0, "ch-work-quiet");
-        assert!(last.1.is_none(), "empty channel has no last_message");
-    }
-
-    #[test]
-    fn channel_previews_excludes_channels_user_is_not_in() {
-        let conn = db();
-        setup(&conn);
-
-        // Carol has her own group — bob is not a member
-        conn.execute("INSERT INTO users (id, email, username) VALUES ('carol', 'carol@x.com', 'carol')", []).unwrap();
-        conn.execute("INSERT INTO groups (id, name, owner_id) VALUES ('g-secret', 'secret', 'carol')", []).unwrap();
-        conn.execute("INSERT INTO group_member (group_id, user_id) VALUES ('g-secret', 'carol')", []).unwrap();
-        conn.execute("INSERT INTO channels (id, group_id, name) VALUES ('ch-secret', 'g-secret', 'private')", []).unwrap();
-
-        let mut stmt = conn.prepare(super::QUERY_CHANNEL_PREVIEWS).unwrap();
-        let channel_ids: Vec<String> = stmt.query_map(
-            rusqlite::params!["bob"],
-            |row| row.get(2),
-        ).unwrap().map(|r| r.unwrap()).collect();
-
-        assert!(!channel_ids.contains(&"ch-secret".to_string()));
-    }
-}
